requirements.txt
dill==0.3.5.1
dnspython==2.2.1
fastapi==0.78.0
httptools==0.5.0
imblearn==0.0
mypy-boto3-s3==1.24.76
pip-chill==1.0.1
pymongo==4.2.0
python-dotenv==0.21.0   #this function is needed in order to read the .env file & make it available as the environment variable
types-s3transfer==0.6.0.post4
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
xgboost==1.6.2
pandas
PyYAML
numpy
scikit-learn
apache-airflow
-e . 
#to trigger the 'setup.py' file, this line of code is needed. Idea behind this is that, when requirements.txt is called, it will automatically call setup.py file & convert our entire source code into library format w/o explicitly need of calling that
# . means current directory & e means editable installation, we are editing setup.py